backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.28.0
google-cloud-firestore==2.19.0
anthropic>=0.40.0
